+++ frontend/src/views/User.vue
@@ -1,5 +1,6 @@
 <script setup>
 import { useScopedI18n } from '@/i18n/app'
+import { useIsMobile } from '../utils/composables'
 
 import { useGlobalState } from '../store'
 
@@ -10,17 +11,19 @@ import BindAddress from './user/BindAddress.vue';
 import UserMailBox from './user/UserMailBox.vue';
 
 const {
-    userTab, globalTabplacement, userSettings
+    userTab, userSettings
 } = useGlobalState()
 
 const { t } = useScopedI18n('views.User')
+const isMobile = useIsMobile()
 
 </script>
 
 <template>
-    <div>
+    <div class="app-shell">
         <UserBar />
-        <n-tabs v-if="userSettings.user_email" type="card" v-model:value="userTab" :placement="globalTabplacement">
+        <n-tabs v-if="userSettings.user_email" class="app-sidebar-tabs" type="card" v-model:value="userTab"
+            :placement="isMobile ? 'top' : 'left'">
             <n-tab-pane name="address_management" :tab="t('address_management')">
                 <AddressMangement />
             </n-tab-pane>